requirements.txt
django-colorfield==0.1.15
Django==2.1.2
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.5